Morning warmup, added service layer checks, and refactored Order Details to be keyless

+++ Jikan/JikanAPI/JikanAPI/Service/JikanService.cs
@@ -1,4 +1,5 @@
 ﻿using JikanAPI.Controllers;
+using JikanAPI.Exceptions;
 using JikanAPI.Models;
 using JikanAPI.Repos;
 using System;
@@ -17,16 +18,27 @@ namespace JikanAPI.Service
         }
         public int AddWatch(Watch toAdd)
         {
+            if (toAdd == null)
+                throw new ArgumentNullException("Cannot create a null watch.");
+
             return _watchRepo.AddWatch(toAdd);
         }
 
         public Watch GetWatchById(int id)
         {
+            if (id <= 0)
+                throw new InvalidIdException("Invalid Id, cannot be <= 0.");
+
             return _watchRepo.GetWatchById(id);
         }
 
         public Watch GetWatchByName(string name)
         {
+            if (name == null)
+                throw new ArgumentNullException("Cannot search by null name.");
+            if (name == "" || name.Length > 50 || name.Trim().Length == 0)
+                throw new InvalidNameException("Invalid name, cannot be empty, white spaces, or too long.");
+            
             return _watchRepo.GetWatchByName(name);
         }
 
@@ -37,31 +49,48 @@ namespace JikanAPI.Service
 
         public List<Watch> GetWatchesByType(string type)
         {
+            if (type == null)
+                throw new ArgumentNullException("Cannot search by null type.");
+            if (type == "" || type.Length > 50 || type.Trim().Length == 0)
+                throw new InvalidNameException("Invalid type, cannot be empty, white spaces, or too long.");
+
             return _watchRepo.GetWatchesByType(type);
         }
 
-        public List<Watch> GetWatchesByPrice(decimal min, decimal max)
+        public List<Watch> GetWatchesByPrice(decimal max)
         {
-            return _watchRepo.GetWatchesByPrice(min, max);
+            return _watchRepo.GetWatchesByPrice(max);
         }
 
         public void EditWatch(Watch toEdit)
         {
+            if (toEdit == null)
+                throw new ArgumentNullException("Cannot edit a null watch.");
+
             _watchRepo.EditWatch(toEdit);
         }
 
         public void DeleteWatch(int id)
         {
+            if (id <= 0)
+                throw new InvalidIdException("Invalid Id, cannot be <= 0.");
+
             _watchRepo.DeleteWatch(id);
         }
 
         public int AddOrder(Order toAdd)
         {
+            if (toAdd == null)
+                throw new ArgumentNullException("Cannot add a null order.");
+
            return _orderRepo.AddOrder(toAdd);
         }
 
         public Order GetOrderById(int id)
         {
+            if (id <= 0)
+                throw new InvalidIdException("Invalid Id, cannot be <= 0.");
+
             return _orderRepo.GetOrderById(id);
         }
 
@@ -72,6 +101,9 @@ namespace JikanAPI.Service
 
         public void DeleteOrder(int id)
         {
+            if (id <= 0)
+                throw new InvalidIdException("Invalid Id, cannot be <= 0.");
+
             _orderRepo.DeleteOrder(id);
         }
